Move spotdl.py to main.py, update tests
