+++ test/loader.py
@@ -1,6 +1,7 @@
 from spotdl import const
 from spotdl import handle
-from spotdl import spotdl
+from spotdl import main
+
 import pytest
 
 
@@ -9,6 +10,7 @@ def load_defaults():
     const.args.overwrite = 'skip'
     const.args.log_level = 10
 
-    spotdl.args = const.args
-    spotdl.log = const.logzero.setup_logger(formatter=const._formatter,
-                                      level=const.args.log_level)
+    main.args = const.args
+    main.log = const.logzero.setup_logger(
+        formatter=const._formatter, level=const.args.log_level
+    )
